Optimize create game
Before this to get list of player and player id we generate the data
twice, so we loop the data twice to get separated name and id.
If user dont have many player data maybe it not become issue, but if
player data is big loop somany time can be issue.
Instead looping twice its better to create the other List variable to
get List of name and id so we just need 1 loop.

Also create dialog in otheer function so the code more clean, why we
produce the same function multiple time if we can combine that.

Signed-off-by: rizky Kharisma <ngeng.ngengs@gmail.com>